+++ Source/Menus/LoadingScreen.cpp
@@ -15,12 +15,15 @@
 
 #include "raylib/raylib.h"
 #include "raylib/rlgl.h"
+#include "RenderTarget.h"
 
 using namespace RTE;
 
+LoadingScreen::LoadingScreen()  { Clear(); }
 void LoadingScreen::Clear() {
 	m_LoadingLogWriter = nullptr;
 	m_LoadingSplashBitmap = nullptr;
+	m_LoadingBackground.reset();
 	m_ProgressListboxBitmap = nullptr;
 	m_ProgressListboxPosX = 0;
 	m_ProgressListboxPosY = 0;
@@ -69,13 +72,21 @@ void LoadingScreen::CreateLoadingSplash(int xOffset) {
 	m_LoadingSplashBitmap = create_bitmap_ex(FrameMan::c_BPP, backbuffer->w, backbuffer->h);
 	clear_bitmap(m_LoadingSplashBitmap);
 
-	StaticSceneLayer loadingSplash;
-	loadingSplash.Create(ContentFile("Base.rte/GUIs/Title/LoadingSplash.png").GetAsBitmap(COLORCONV_NONE, false), false, Vector(), true, false, Vector(1.0F, 0));
-	loadingSplash.SetOffset(Vector(static_cast<float>(((loadingSplash.GetBitmap()->w - g_WindowMan.GetResX()) / 2) + xOffset), 0));
+	m_LoadingBackground = std::make_unique<StaticSceneLayer>();
+	m_LoadingBackground->Create(ContentFile("Base.rte/GUIs/Title/LoadingSplash.png").GetAsBitmap(COLORCONV_NONE, false), false, Vector(), true, false, Vector(1.0F, 0));
+	m_LoadingBackground->SetOffset(Vector(static_cast<float>(((m_LoadingBackground->GetBitmap()->w - g_WindowMan.GetResX()) / 2) + xOffset), 0));
 
-	Box loadingSplashTargetBox(Vector(0, static_cast<float>((g_WindowMan.GetResY() - loadingSplash.GetBitmap()->h) / 2)), static_cast<float>(g_WindowMan.GetResX()), static_cast<float>(loadingSplash.GetBitmap()->h));
+	Box loadingSplashTargetBox(Vector(0, static_cast<float>((g_WindowMan.GetResY() - m_LoadingBackground->GetBitmap()->h) / 2)), static_cast<float>(g_WindowMan.GetResX()), static_cast<float>(m_LoadingBackground->GetBitmap()->h));
+	RenderTarget defaultTarget{
+		FloatRect(0, 0, g_WindowMan.GetResX(), g_WindowMan.GetResY()),
+		FloatRect(0, 0, g_WindowMan.GetResX(), g_WindowMan.GetResY()),
+		0,
+		Texture2D(),
+		true
+	};
+	defaultTarget.Begin();
 	g_WindowMan.ClearBackbuffer();
-	loadingSplash.Draw(loadingSplashTargetBox, loadingSplashTargetBox);
+	m_LoadingBackground->Draw(loadingSplashTargetBox, loadingSplashTargetBox);
 	rlDrawRenderBatchActive();
 	g_WindowMan.Present();
 }
@@ -140,7 +151,10 @@ void LoadingScreen::LoadingSplashProgressReport(const std::string& reportString,
 
 		blit(g_LoadingScreen.m_ProgressListboxBitmap, g_FrameMan.GetBackBuffer32(), 0, 0, g_LoadingScreen.m_ProgressListboxPosX, g_LoadingScreen.m_ProgressListboxPosY, g_LoadingScreen.m_ProgressListboxBitmap->w, g_LoadingScreen.m_ProgressListboxBitmap->h);
 
+		Box loadingSplashTargetBox(Vector(0, static_cast<float>((g_WindowMan.GetResY() - g_LoadingScreen.m_LoadingBackground->GetBitmap()->h) / 2)), static_cast<float>(g_WindowMan.GetResX()), static_cast<float>(g_LoadingScreen.m_LoadingBackground->GetBitmap()->h));
 		g_WindowMan.ClearBackbuffer(false);
+		g_WindowMan.GetScreenBuffer()->Begin();
+		g_LoadingScreen.m_LoadingBackground->Draw(loadingSplashTargetBox, loadingSplashTargetBox);
 		g_WindowMan.UploadFrame();
 	}
 }